Various minor fixes and improvements from beebem-mac

+++ Src/PALRom.cpp
@@ -223,7 +223,7 @@ PALRomType GuessRomType(uint8_t *Rom, uint32_t Size)
     char RomName[11];
     memcpy(RomName, &Rom[9], 10);
     RomName[10] = '\0';
-    int Crc = crc32(0, Rom, Size);
+    int Crc = (int)crc32(0, Rom, Size);
 
     if (strstr(RomName, "ViewSpell") && Size == PALROM_64K && Crc == 0xE56B0E01)
     {
